Rename file and type
