+++ src/bug-filing/azure-boards/create-azure-boards-issue-filing-url.ts
@@ -3,10 +3,10 @@
 import { EnvironmentInfo } from '../../common/environment-info-provider';
 import { CreateIssueDetailsTextData } from '../../common/types/create-issue-details-text-data';
 import { title } from '../../content/strings/application';
+import { createIssueDetailsBuilder } from '../common/create-issue-details-builder';
 import { HTMLFactory } from '../common/html-factory';
 import { HTTPQueryBuilder } from '../common/http-query-builder';
-import { createIssueDetailsBuilder } from '../common/issue-details-builder';
-import { IssueDetailsGetter } from '../common/issue-details-getter';
+import { IssueDetailsBuilder } from '../common/issue-details-builder';
 import { IssueFilingUrlStringUtils, IssueUrlCreationUtils } from '../common/issue-filing-url-string-utils';
 import { AzureBoardsBugFilingSettings } from './azure-boards-bug-filing-service';
 
@@ -17,7 +17,7 @@ const buildTags = (createBugData: CreateIssueDetailsTextData, standardTags: stri
 
 export const createAzureBoardsIssueFilingUrlProvider = (
     stringUtils: IssueUrlCreationUtils,
-    issueDetailsGetter: IssueDetailsGetter,
+    issueDetailsGetter: IssueDetailsBuilder,
     queryBuilderProvider: () => HTTPQueryBuilder,
 ) => {
     return (settingsData: AzureBoardsBugFilingSettings, bugData: CreateIssueDetailsTextData, environmentInfo: EnvironmentInfo) => {
