+++ src/tests/unit/tests/bug-filing/common/issue-details-builder.test.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 import { It, Mock, MockBehavior } from 'typemoq';
-import { createIssueDetailsBuilder } from '../../../../../bug-filing/common/issue-details-builder';
+import { createIssueDetailsBuilder } from '../../../../../bug-filing/common/create-issue-details-builder';
 import { MarkupFactory } from '../../../../../bug-filing/common/markup-factory';
 import { EnvironmentInfo } from '../../../../../common/environment-info-provider';
