+++ src/tests/unit/tests/bug-filing/github/create-github-bug-filing-url.test.ts
@@ -2,7 +2,7 @@
 // Licensed under the MIT License.
 import { IMock, Mock } from 'typemoq';
 
-import { IssueDetailsGetter } from '../../../../../bug-filing/common/issue-details-getter';
+import { IssueDetailsBuilder } from '../../../../../bug-filing/common/issue-details-builder';
 import { createGitHubIssueFilingUrlProvider } from '../../../../../bug-filing/github/create-github-bug-filing-url';
 import { IssueFilingUrlProvider } from '../../../../../bug-filing/types/bug-filing-service';
 import { IssueUrlCreationUtils } from './../../../../../bug-filing/common/issue-filing-url-string-utils';
@@ -14,7 +14,7 @@ describe('createGitHubBugFilingUrlTest', () => {
     let sampleIssueDetailsData;
     let settingsData: GitHubBugFilingSettings;
     let stringUtilsMock: IMock<IssueUrlCreationUtils>;
-    let issueDetailsGetter: IMock<IssueDetailsGetter>;
+    let issueDetailsGetter: IMock<IssueDetailsBuilder>;
     let testObject: IssueFilingUrlProvider<GitHubBugFilingSettings>;
 
     beforeEach(() => {
@@ -46,7 +46,7 @@ describe('createGitHubBugFilingUrlTest', () => {
         stringUtilsMock.setup(utils => utils.getTitle(sampleIssueDetailsData)).returns(() => 'test title');
         stringUtilsMock.setup(utils => utils.appendSuffixToUrl(settingsData.repository, 'issues')).returns(() => 'test appendSuffixToUrl');
 
-        issueDetailsGetter = Mock.ofType<IssueDetailsGetter>();
+        issueDetailsGetter = Mock.ofType<IssueDetailsBuilder>();
         issueDetailsGetter.setup(getter => getter(environmentInfo, sampleIssueDetailsData)).returns(() => 'test issue details');
         testObject = createGitHubIssueFilingUrlProvider(stringUtilsMock.object, issueDetailsGetter.object);
     });
